Add some cast to avoid platform dependant warnings.
git-svn-id: file:///home/yoann/SVN/trunk/libprelude@7184 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/idmef-criterion-value.c
@@ -112,7 +112,7 @@ static int btime_parse_month(const char *value, int *out)
                 "december"  
         };
 
-        if ( isdigit(*value) ) {
+        if ( isdigit((int) *value) ) {
                 *out = atoi(value) - 1;
                 return 0;
         }
@@ -142,7 +142,7 @@ static int btime_parse_wday(const char *value, int *out)
                 "saturday",
         };
         
-        if ( isdigit(*value) ) {
+        if ( isdigit((int) *value) ) {
                 *out = atoi(value) - 1;
                 return 0;
         }
